WIP
Because we now have an interface for 0.0.0.0/0[...] shunts we get a
default route in table 220.  This causes all sorts of problems if we
don't have shunts for other subnets/routes.  An example are the shunts
for port 22 in the swanctl/manual-prio scenario.  Before, these got us no
interface (0.0.0.0 was the local and src IP, which let get_interface_name
fail) so no route was installed.
Since routes for shunts with protocl/ports are tricky anyway (they are
too broad) we might want to ignore these.